Fix name 'WriteOBJ' is not defined
